docs: listnodes option_will_fund lease fields are u16
The lease_rates wire fields are u16 (wire/peer_wire.csv:215-217)
and json_add_lease_rates serializes them from the u16 struct
members (common/json_stream.c:671-683).  Match the openchannel2
hook response types.

Changelog-None

+++ cln-rpc/src/model.rs
@@ -7203,11 +7203,11 @@ pub mod responses {
 	#[derive(Clone, Debug, Deserialize, Serialize)]
 	pub struct ListnodesNodesOptionWillFund {
 	    pub channel_fee_max_base_msat: Amount,
-	    pub channel_fee_max_proportional_thousandths: u32,
+	    pub channel_fee_max_proportional_thousandths: u16,
 	    pub compact_lease: String,
-	    pub funding_weight: u32,
+	    pub funding_weight: u16,
 	    pub lease_fee_base_msat: Amount,
-	    pub lease_fee_basis: u32,
+	    pub lease_fee_basis: u16,
 	}
 
 	/// ['Type of connection (until 23.08, `websocket` was also allowed).']
